requirements.txt
python-dotenv==0.18.0
argparse==1.4.0
requests==2.25.1
